refactor(feat/meetup): 모임 데이터 요청 함수 분리 및 이름 변경
